requirements.txt
cchardet==1.1.1
decorator==4.0.10
docutils==0.13.1
networkx==1.11
nltk==3.2.1
numpy==1.11.3
python-cjson==1.1.0
PyYAML==3.12
scipy==0.18.1
statistics==1.0.3.5
